split up keys tests
