+++ lib/keys/index.js
@@ -14,8 +14,12 @@ exports.properties = {
     if (val !== this.sort) {
       this.sort = val
       if (this._keys) {
-        this._keys._ = new Array(this._keys.length)
-        sort(this, this._keys, val)
+        if (val === null) {
+          console.log('REMOVE SORT')
+        } else {
+          this._keys._ = new Array(this._keys.length)
+          sort(this, this._keys, val)
+        }
       }
     }
   }
@@ -34,12 +38,11 @@ exports.define = {
   keys,
   addKey (key, target) {
     var keys = this._keys
+    var copy
     if (!this._ownKeys) {
-      // order will add complications here -- maybe use a filter when you use order?
+      copy = true
       keys = this._keys = keys ? keys.concat([ key ]) : [ key ]
-      if (this._filters) {
-        addFilters(this, key, true)
-      } else {
+      if (!this._filters) {
         this._filters = false
       }
       this._ownKeys = true
@@ -49,9 +52,6 @@ exports.define = {
       } else {
         keys.push(key)
       }
-      if (this._filters) {
-        addFilters(this, key)
-      }
     }
     if ('sort' in this) {
       const sort = this.sort
@@ -60,6 +60,9 @@ exports.define = {
       }
     }
     instances(this, target, this.addKey, keys, key)
+    if (this._filters) {
+      addFilters(this, key, copy)
+    }
   },
   removeKey (key, target) {
     var keys = this._keys
@@ -74,7 +77,7 @@ exports.define = {
           for (let i = 0; i < len; i++) {
             if (keys[i] === key) {
               keys.splice(i, 1)
-              if (keys._) { keys._.splice(i, 1) }
+              if ('_' in keys) { keys._.splice(i, 1) }
               break
             }
           }
@@ -97,7 +100,7 @@ exports.define = {
         for (let i = 0; i < len; i++) {
           if (keys[i] === key) {
             keys.splice(i, 1)
-            if (keys._) { keys._.splice(i, 1) }
+            if ('_' in keys) { keys._.splice(i, 1) }
             break
           }
         }
@@ -109,7 +112,6 @@ exports.define = {
     instances(this, target, this.removeKey, keys, key)
   },
   removeAllKeys () {
-    // removeAllKeys will not update instances
     if (!this._ownKeys) {
       this._ownKeys = true
     }
