+++ lib/keys/sort/quick.js
@@ -1,9 +1,8 @@
 'use strict'
 module.exports = function quickSort (target, keys, left, right, field, orders) {
-  var pivot, index
   if (left < right) {
-    pivot = right
-    index = partition(target, keys, pivot, left, right, field, orders)
+    const pivot = right
+    const index = partition(target, keys, pivot, left, right, field, orders)
     quickSort(target, keys, left, index - 1, field, orders)
     quickSort(target, keys, index + 1, right, field, orders)
   }
