+++ test/keys/basic.js
@@ -1,0 +1,24 @@
+'use strict'
+const test = require('tape')
+const Base = require('../../')
+const isEmpty = require('vigour-util/is/empty')
+test('keys', function (t) {
+  const base = new Base({
+    a: true,
+    b: true,
+    val: 'something'
+  })
+  t.equal(base.keys().length, 2, 'correct length')
+  base.removeProperty(base.a, 'a')
+  t.equal(base.keys().length, 1, 'correct length after removal')
+  base.setKey('c', true)
+  t.equal(base.keys().length, 2, 'correct length after setKey')
+  base.reset()
+  t.equal(isEmpty(base), true, 'empty after reset')
+  t.equal(base.keys().length, 0, 'keys are false after reset')
+  base.set({ d: true })
+  t.equal(base.keys().length, 1, 'correct length after set')
+  base.set({ c: null })
+  t.equal(base.keys().length, 1, 'correct length after set with null')
+  t.end()
+})
