+++ test/keys/index.js
@@ -1,0 +1,5 @@
+'use strict'
+require('./basic')
+require('./inheritance')
+require('./order')
+require('./filters')
